// yourcode/webapp/src/app/shared/models/filme.ts
import { Rating } from './rating';

export class Filme {
  Title: string;
  Year: string;
  Rated: string;
  Released: string;
  Runtime: string;
  Genre: string;
  Director: string;
  Writer: string;
  Actors: string;
  Plot: string;
  Language: string;
  Country: string;
  Awards: string;
  Poster: string;
  Ratings: Rating[];
  Metascore: string;
  imdbRating: string;
  imdbVotes: string;
  imdbID: string;
  Type: string;
  Dvd: string;
  BoxOffice: string;
  Production: string;
  Website: string;

  // prop front
  favorito: boolean;
}